requirements.txt
torch>=2.1
PyYAML>=6.0
numpy>=1.24
mpi4py>=3.1
pandas>=2.0
pyarrow>=12.0
matplotlib>=3.7
